Lots of code fixes
