--- LibraryWebApp/Views/Account/Details.cshtml
@@ -1,74 +1,0 @@
-﻿@using Microsoft.AspNetCore.Identity
-@inject UserManager<ApplicationUser> UserManager
-@model IEnumerable<RentedBook>
-@{
-    ViewData["Title"] = "My profile";
-}
-<h1>Rented</h1>
-
-<table class="table">
-    <thead>
-        <tr>
-            <th>
-                <label class="control-label">Title</label>
-            </th>
-            <th>
-                <label class="control-label">Rented Date</label>
-            </th>
-            <th>
-                <label class="control-label">Return Date</label>
-            </th>
-            <th></th>
-        </tr>
-    </thead>
-    <tbody>
-        @foreach (var book in Model)
-        {
-            <tr>
-                <td>
-                    @book.Book.Title
-                </td>
-                <td>
-                    @book.RentalDate.ToString("dd/MM/yyyy")
-                </td>
-                <td>
-                    @book.ReturnDate
-                </td>
-                <td>
-                    @if (book.ReturnedAt.HasValue)
-                    {
-                        if (book.ReturnedAt <= book.ReturnDate)
-                        {
-                            <p>✔ Returned on time</p>
-                        }
-                        else
-                        {
-                            <p>✘ Returned late</p>
-                        }
-                    }
-                    else
-                    {  
-                            if (DateOnly.FromDateTime(DateTime.Now) <= book.ReturnDate)
-                            {
-                                <form asp-controller="Account" asp-action="Return">
-                                    <input type="hidden" name="Id" value="@book.Id" />
-                                    <button type="submit" class="btn btn-primary">Return Book</button>
-                                </form>
-                            }
-                            else
-                            {
-                                <form asp-controller="Account" asp-action="Return">
-                                    <input type="hidden" name="Id" value="@book.Id" />
-                                    <button type="submit" class="btn btn-primary">Return Book Late</button>
-                                </form>
-                            }
-                    }
-                </td>
-            </tr>
-        }
-    </tbody>
-</table>
-
-@section Scripts {
-    <partial name="_ValidationScriptsPartial" />
-}
