+++ LibraryWebApp/Views/Book/Index.cshtml
@@ -36,7 +36,7 @@
                     @book.Title
                 </td>
                 <td>
-                    @book.PublicationDate.ToString("MM/dd/yyyy")
+                    @book.PublicationDate.ToString(Globals.DateFormat)
                 </td>
                 <td>
                     @book.AvailableCount
